GOO files decoding

+++ src/test/java/futurelink/msla/formats/creality/CXDLPFileTest.java
@@ -18,7 +18,7 @@ public class CXDLPFileTest extends CommonTestRoutines {
         delete_file(outFile); // Clean up files just in case
 
         var file = (CXDLPFile) FileFactory.instance.create("CREALITY HALOT-ONE PLUS");
-        file.options().set("BottomExposureTime", 12);
+        file.options().set("BottomExposureTime", (short) 12);
 
         var pngFileLayers = new String[]{
                 resourceFile("test_data/CXDLPFileTest/Layer_1.png"),
